Made a new RMD, added fips codes
I want to also consider population in this project. The PLACES dataset does not have population data, so I retrieved population data from the ACS and joined by fips code. I created a second rmd file so that we don't have to keep using the file that has code we should no longer use but that is relevant to our project.