Added ggplot and images
I grouped the data by state and found the average rate of illnesses by state then graphed them. I then added those images to the images folder so they can later be used in a QMD file. I don't want to repetitively perform computation with heavy loads in an online Rmd file. Am willing to change if the professor says we have to.